--- resources/web.qrc
@@ -1,9 +1,0 @@
-<RCC>
-  <qresource prefix="/review_heatmap">
-    <file>web/cal-heatmap.css</file>
-    <file>web/cal-heatmap.js</file>
-    <file>web/d3.min.js</file>
-    <file>web/review-heatmap.js</file>
-    <file>web/review-heatmap.css</file>
-  </qresource>
-</RCC>
