+++ src/web/css/review-heatmap.css
